+++ src/stam_annotator/stam_fetcher/utility.py
@@ -1,10 +1,9 @@
 import json
-import subprocess
 from pathlib import Path
 
 from github import Github, GithubException
 
-from stam_annotator.exceptions import RepoCloneError, RepoDoesNotExist
+from stam_annotator.exceptions import RepoDoesNotExist
 
 
 def check_repo_exists(token, org_name, repo_name):
@@ -14,19 +13,6 @@ def check_repo_exists(token, org_name, repo_name):
         org.get_repo(repo_name)
     except GithubException:
         raise RepoDoesNotExist(org_name, repo_name)
-
-
-def clone_repo(org, repo_name, token, destination_folder: Path):
-    try:
-        """make a inner folder with source org name and clone the repo in it"""
-        repo_url = f"https://{token}@github.com/{org}/{repo_name}.git"
-        subprocess.run(["git", "clone", repo_url, destination_folder], check=True)
-        print(
-            f"[SUCCESS]: Repository {repo_name} cloned successfully into {destination_folder}"
-        )
-
-    except subprocess.CalledProcessError as e:
-        raise RepoCloneError(org, repo_name, e)
 
 
 def add_base_path_to_stam_annotation_files(base_path: Path):
