+++ src/stam_annotator/stam_manager.py
@@ -7,7 +7,9 @@ from stam import AnnotationDataSet, Annotations, AnnotationStore
 
 from stam_annotator.config import AnnotationEnum, AnnotationGroupEnum
 from stam_annotator.loaders.opa_loader import OpaAnnotation
-from stam_annotator.utility import convert_opf_stam_annotation_to_dictionary
+from stam_annotator.to_stam_convertor.utility import (
+    convert_opf_stam_annotation_to_dictionary,
+)
 
 
 def load_stam_from_json(file_path: Union[str, Path]) -> AnnotationStore:
