+++ src/stam_annotator/to_stam_convertor/opf_to_stam.py
@@ -12,7 +12,7 @@ from stam_annotator.loaders.annotation_store import (
     convert_opf_for_pre_stam_format,
 )
 from stam_annotator.loaders.opf_loader import create_opf_annotation_instance
-from stam_annotator.utility import load_opf_annotations_from_yaml
+from stam_annotator.loaders.utility import load_opf_annotations_from_yaml
 
 
 def get_uuid():
